Fixed select & popupButton styling
