+++ src/renderer/components/users/UserPoup.tsx
@@ -13,6 +13,7 @@ import {
   PopupDialogTabPage,
   PopupButtonWrapper
 } from '../generic-styled-components/popupDialog';
+import { SelectWrap } from '../generic-styled-components/Select'
 import { FaTimes } from 'react-icons/fa';
 import { getPhrase } from '@/renderer/helpers/lang';
 import { Button } from '../generic-styled-components/button';
@@ -123,13 +124,15 @@ export const UserPopup = (props: IProps) => {
                 <PopupDialogInputName>
                   {getPhrase('user_popup_permissions')}
                 </PopupDialogInputName>
-                <Select
-                  value={user
-                    .getPermissionStrings()
-                    .map(item => ({ label: item, value: item }))}
-                  isDisabled={true}
-                  isMulti={true}
-                />
+                <SelectWrap width={'100%'}>
+                  <Select
+                    value={user
+                      .getPermissionStrings()
+                      .map(item => ({ label: item, value: item }))}
+                    isDisabled={true}
+                    isMulti={true}
+                  />
+                </SelectWrap>
               </PopupDialogInputWrapper>
               <PopupDialogInputWrapper>
                 <PopupDialogInputName>
